Move the lane's tile math out of the pixi-importing module
CI runs Node 20, which has no global `navigator`; pixi.js touches it at
import time, so terrainAlignment.test.ts failed at import as soon as it
reached utils/terrainLayer.ts. It passed locally only because Node 22
defines that global.

The coordinate and tile-index helpers were never rendering code, so they
move to utils/laneGrid.ts, which imports no pixi. terrainLayer re-exports
gameToPixel/pixelToGame/TILE_RADIUS_SCALE, so every existing caller is
untouched, and the test imports the helpers the renderer actually uses
rather than restating the rule.
